requirements.txt
certifi==2019.11.28
chardet==3.0.4
click==7.1.1
docutils==0.16
Flask==1.1.2
Flask-Cors==3.0.8
Flask-HTTPAuth==4.0.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.11.2
lockfile==0.12.2
MarkupSafe==1.1.1
peewee==3.13.3
psutil==5.7.0
PyMySQL==0.9.3
python-daemon==2.2.4
python-dateutil==2.8.1
requests==2.22.0
six==1.14.0
urllib3==1.25.8
Werkzeug==1.0.1
